rotate: really fix overflow

+++ modules/video_filter/rotate.c
@@ -95,8 +95,8 @@ static void store_trigo( struct filter_sys_t *sys, float f_angle )
 {
     f_angle *= (float)(M_PI / 180.); /* degrees -> radians */
 
-    uint16_t i_sin = lroundf(sinf(f_angle) * 4096.f);
-    uint16_t i_cos = lroundf(cosf(f_angle) * 4096.f);
+    unsigned i_sin = lroundf(sinf(f_angle) * 4096.f);
+    unsigned i_cos = lroundf(cosf(f_angle) * 4096.f);
     atomic_store( &sys->sincos, (i_cos << 16u) | (i_sin << 0u));
 }
